Replace freeze API with DisableInput improved resource
ActionState::freeze was introduced to consume actions. But it turns
out not very ergonomic. Also freeze require to spawn an entity or
init the resource should which is not always the case.
In this iteration I used run criteria to make the code nicer.
Also I extended disable_input test to check if global and entity input
is released on disable.